requirements.txt
streamlit
yfinance
pandas
plotly.express
plotly.graph_objects